+++ dutchiepay/src/app/(modals)/change-number/page.jsx
@@ -1,15 +1,14 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
 import PhoneAuth from '@/app/_components/_user/_phone/PhoneAuth';
+import ProtectedRoute from '@/app/_components/ProtectedRoute';
 import axios from 'axios';
 import { useForm } from 'react-hook-form';
 import useReissueToken from '@/app/hooks/useReissueToken';
 import { useSelector } from 'react-redux';
+import { useState } from 'react';
 
 export default function ChangeNumber() {
-  const isLoggedIn = useSelector((state) => state.login.isLoggedIn);
   const access = useSelector((state) => state.login.access);
   const [isPhoneAuth, setIsPhoneAuth] = useState(false); // 핸드폰 인증 요청 여부
   const [isCodeMatch, setIsCodeMatch] = useState(null);
@@ -26,13 +25,6 @@ export default function ChangeNumber() {
     reValidateMode: 'onblur',
     shouldFocusError: true,
   });
-
-  useEffect(() => {
-    if (!isLoggedIn || !access) {
-      alert('비정상적인 접속');
-      closeWindow();
-    }
-  }, [isLoggedIn, access]);
 
   const onSubmit = async (formData) => {
     try {
@@ -73,57 +65,56 @@ export default function ChangeNumber() {
     }
   };
 
-  const closeWindow = () => {
-    window.close();
-  };
-
   return (
-    <section className="min-w-[620px] p-[32px]">
-      <h1 className="text-3xl font-bold">휴대폰 번호 변경</h1>
-      <p className="text-xs text-gray--500">
-        휴대폰 번호 변경을 위해 휴대폰 본인 인증을 필요로 합니다.
-        <br />
-        변경하실 휴대폰 번호를 입력 후, 인증하기를 눌러주세요.
-      </p>
-      <form
-        className="mt-[80px] w-[450px] mx-auto"
-        onSubmit={handleSubmit(onSubmit)}
-      >
-        <PhoneAuth
-          register={register}
-          watch={watch}
-          errors={errors}
-          touchedFields={touchedFields}
-          setValue={setValue}
-          isPhoneAuth={isPhoneAuth}
-          setIsPhoneAuth={setIsPhoneAuth}
-          isCodeMatch={isCodeMatch}
-          setIsCodeMatch={setIsCodeMatch}
-        />
-        {isCodeMatch && (
-          <p className="text-sm">
-            휴대폰 번호 인증이 완료되었습니다.
-            <br />
-            <strong>변경</strong> 버튼을 누르시면 정상적으로 번호가 변경됩니다.
-          </p>
-        )}
-        <div className="mt-[40px] flex gap-[12px] justify-center">
-          <button
-            className={`text-red-500 text-sm rounded-lg px-[24px] py-[8px] ${!isCodeMatch ? 'cursor-not-allowed bg-gray--100 text-white' : 'bg-red--100'}`}
-            type="submit"
-            disabled={!isCodeMatch}
-          >
-            변경
-          </button>
-          <button
-            className="text-blue--500 text-sm border border-blue--200 rounded-lg px-[24px] py-[8px]"
-            type="button"
-            onClick={closeWindow}
-          >
-            취소
-          </button>
-        </div>
-      </form>
-    </section>
+    <ProtectedRoute>
+      <section className="min-w-[620px] p-[32px]">
+        <h1 className="text-3xl font-bold">휴대폰 번호 변경</h1>
+        <p className="text-xs text-gray--500">
+          휴대폰 번호 변경을 위해 휴대폰 본인 인증을 필요로 합니다.
+          <br />
+          변경하실 휴대폰 번호를 입력 후, 인증하기를 눌러주세요.
+        </p>
+        <form
+          className="mt-[80px] w-[450px] mx-auto"
+          onSubmit={handleSubmit(onSubmit)}
+        >
+          <PhoneAuth
+            register={register}
+            watch={watch}
+            errors={errors}
+            touchedFields={touchedFields}
+            setValue={setValue}
+            isPhoneAuth={isPhoneAuth}
+            setIsPhoneAuth={setIsPhoneAuth}
+            isCodeMatch={isCodeMatch}
+            setIsCodeMatch={setIsCodeMatch}
+          />
+          {isCodeMatch && (
+            <p className="text-sm">
+              휴대폰 번호 인증이 완료되었습니다.
+              <br />
+              <strong>변경</strong> 버튼을 누르시면 정상적으로 번호가
+              변경됩니다.
+            </p>
+          )}
+          <div className="mt-[40px] flex gap-[12px] justify-center">
+            <button
+              className={`text-red-500 text-sm rounded-lg px-[24px] py-[8px] ${!isCodeMatch ? 'cursor-not-allowed bg-gray--100 text-white' : 'bg-red--100'}`}
+              type="submit"
+              disabled={!isCodeMatch}
+            >
+              변경
+            </button>
+            <button
+              className="text-blue--500 text-sm border border-blue--200 rounded-lg px-[24px] py-[8px]"
+              type="button"
+              onClick={() => window.close()}
+            >
+              취소
+            </button>
+          </div>
+        </form>
+      </section>
+    </ProtectedRoute>
   );
 }
