+++ dutchiepay/src/app/(modals)/withdraw-auth/page.jsx
@@ -1,15 +1,14 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
 import PhoneAuth from '@/app/_components/_user/_phone/PhoneAuth';
+import ProtectedRoute from '@/app/_components/ProtectedRoute';
 import axios from 'axios';
 import { useForm } from 'react-hook-form';
 import useReissueToken from '@/app/hooks/useReissueToken';
 import { useSelector } from 'react-redux';
+import { useState } from 'react';
 
 export default function WithdrawAuth() {
-  const isLoggedIn = useSelector((state) => state.login.isLoggedIn);
   const access = useSelector((state) => state.login.access);
   const [isPhoneAuth, setIsPhoneAuth] = useState(false); // 핸드폰 인증 요청 여부
   const [isCodeMatch, setIsCodeMatch] = useState(null);
@@ -28,12 +27,6 @@ export default function WithdrawAuth() {
     shouldFocusError: true,
   });
 
-  useEffect(() => {
-    if (!isLoggedIn || !access) {
-      alert('비정상적인 접속');
-      closeWindow();
-    }
-  }, [access, isLoggedIn]);
   const fetchDeleteUser = async (accessToken) => {
     const response = await axios.delete(
       loginType === 'email'
@@ -45,7 +38,7 @@ export default function WithdrawAuth() {
         },
       }
     );
-    return response; // 성공적으로 삭제된 경우 응답 반환
+    return response;
   };
 
   const handleWithdraw = async () => {
@@ -58,7 +51,7 @@ export default function WithdrawAuth() {
         fetchDeleteUser(access);
         window.opener.postMessage({ type: 'WITHDRAW' }, window.location.origin);
         alert('정상적으로 탈퇴처리 되었습니다.');
-        closeWindow();
+        window.close();
       } catch (error) {
         if (error.response.data.message === '액세스 토큰이 만료되었습니다.') {
           const reissueResponse = await refreshAccessToken();
@@ -77,54 +70,52 @@ export default function WithdrawAuth() {
     }
   };
 
-  const closeWindow = () => {
-    window.close();
-  };
-
   return (
-    <section className="min-w-[620px] p-[32px]">
-      <h1 className="text-3xl font-bold mb-4">회원탈퇴</h1>
-      <p className="text-xs text-gray--500 mb-6">
-        회원탈퇴를 위해서는 휴대폰 본인 인증을 필요로 합니다.
-        <br />
-        휴대폰 번호를 입력 후, 인증하기를 눌러주세요.
-      </p>
-
-      <PhoneAuth
-        register={register}
-        watch={watch}
-        errors={errors}
-        touchedFields={touchedFields}
-        setValue={setValue}
-        isPhoneAuth={isPhoneAuth}
-        setIsPhoneAuth={setIsPhoneAuth}
-        isCodeMatch={isCodeMatch}
-        setIsCodeMatch={setIsCodeMatch}
-        trigger={trigger}
-      />
-      {isCodeMatch && (
-        <p className="text-sm mt-4">
-          휴대폰 번호 인증이 완료되었습니다.
+    <ProtectedRoute>
+      <section className="min-w-[620px] p-[32px]">
+        <h1 className="text-3xl font-bold mb-4">회원탈퇴</h1>
+        <p className="text-xs text-gray--500 mb-6">
+          회원탈퇴를 위해서는 휴대폰 본인 인증을 필요로 합니다.
           <br />
-          <strong>확인</strong> 버튼을 누르시면 정상적으로 탈퇴가 진행됩니다.
+          휴대폰 번호를 입력 후, 인증하기를 눌러주세요.
         </p>
-      )}
-      <div className="mt-[40px] flex gap-[12px] justify-center">
-        <button
-          className={`text-red-500 text-sm rounded-lg px-[24px] py-[8px] ${!isCodeMatch ? 'cursor-not-allowed bg-gray--100 text-white' : 'bg-red--100'}`}
-          onClick={handleWithdraw}
-          disabled={!isCodeMatch}
-        >
-          탈퇴하기
-        </button>
-        <button
-          className="text-blue--500 text-sm border border-blue--200 rounded-lg px-[24px] py-[8px]"
-          type="button"
-          onClick={closeWindow}
-        >
-          취소
-        </button>
-      </div>
-    </section>
+
+        <PhoneAuth
+          register={register}
+          watch={watch}
+          errors={errors}
+          touchedFields={touchedFields}
+          setValue={setValue}
+          isPhoneAuth={isPhoneAuth}
+          setIsPhoneAuth={setIsPhoneAuth}
+          isCodeMatch={isCodeMatch}
+          setIsCodeMatch={setIsCodeMatch}
+          trigger={trigger}
+        />
+        {isCodeMatch && (
+          <p className="text-sm mt-4">
+            휴대폰 번호 인증이 완료되었습니다.
+            <br />
+            <strong>확인</strong> 버튼을 누르시면 정상적으로 탈퇴가 진행됩니다.
+          </p>
+        )}
+        <div className="mt-[40px] flex gap-[12px] justify-center">
+          <button
+            className={`text-red-500 text-sm rounded-lg px-[24px] py-[8px] ${!isCodeMatch ? 'cursor-not-allowed bg-gray--100 text-white' : 'bg-red--100'}`}
+            onClick={handleWithdraw}
+            disabled={!isCodeMatch}
+          >
+            탈퇴하기
+          </button>
+          <button
+            className="text-blue--500 text-sm border border-blue--200 rounded-lg px-[24px] py-[8px]"
+            type="button"
+            onClick={() => window.close()}
+          >
+            취소
+          </button>
+        </div>
+      </section>
+    </ProtectedRoute>
   );
 }
